Refactor record/playback tests: one class per file under tests/ subfolder
Restructure per repo policy: split the single test module into one test class
per file, moved to tests/record_playback_llm_server/tests/ (separate from app
code). Fixtures live in that subfolder's conftest.py (start_app + the OPENAI-key
autouse override, moved out of the app dir); fake upstream handlers are one
class per file (CountingUpstream/StreamUpstream/RateLimitedUpstream); the former
module-level test helpers become static methods on ProxyTestClient.

Also update the tests to the unified-responses Cassette API (they were failing
on this branch): Cassette now requires use_multi_mode and stores every response
under a 'responses' list. Adjusted constructions/assertions accordingly, renamed
the dedup test to match put_response, and added a single-mode retention test.

23 tests pass; flake8 clean; pylint 10.00/10 (tests dir included).
